Rename transformer package to filters
