+++ nexus/nexus-test/nexus-testsuite-support/src/main/java/org/sonatype/nexus/testsuite/support/filters/CompositeFilter.java
@@ -10,7 +10,7 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.testsuite.support.transformers;
+package org.sonatype.nexus.testsuite.support.filters;
 
 import static com.google.common.base.Preconditions.checkNotNull;
